test: move duplicate-entry
