+++ test/integration/duplicate-entry/fixtures/package.json
