+++ test/integration/duplicate-entry/fixtures/src/index.ts
