+++ test/integration/duplicate-entry/fixtures/tsconfig.json
